requirements.txt
# Web Framework & Server
fastapi==0.103.2
starlette==0.27.0
uvicorn[standard]==0.23.2

# Database & ORM
sqlalchemy==2.0.21
pgvector==0.2.3
psycopg2-binary==2.9.9

# Validation & Settings
pydantic==2.4.2
pydantic-settings==2.0.3

# Templating & Web Utilities
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.0

# Bioinformatics & Data
biopython==1.81
biotite>=0.40.0
pandas==2.1.1
numpy==1.26.0

# CLI Tools
typer==0.9.0
click==8.1.3
rich==13.6.0